Fix code

+++ HomeWork2/HomeWork2.cpp
@@ -9,9 +9,9 @@ int main()
 	int a, b, c;
 	cin >> a >> b >> c;
 
-	int d = (b*b) - (4 * a*c);
-	int x1 = (-b + sqrt(d)) / (2 * a);
-	int x2 = (-b - sqrt(d)) / (2 * a);
+	double d = (b*b) - (4 * a*c);
+	double x1 = (-b + sqrt(d)) / (2.0 * a);
+	double x2 = (-b - sqrt(d)) / (2.0 * a);
 
 	cout << "x1 = " << x1 << endl;
 	cout << "x2 = " << x2 << endl;
